use non-session cookies in bundle with an expiration date
Many browsers do not clean up session cookies when you close them. So the rule of thumb must be: For having a consistent behaviour across all browsers, don't rely solely on browser behaviour for proper clean-up of session cookies. It is safer to use non-session cookies (IsPersistent == true) in bundle with an expiration date.
See http://blog.petersondave.com/cookies/Session-Cookies-in-Chrome-Firefox-and-Sitecore/